+++ src/component/TOS/TOSModal.tsx
@@ -4,11 +4,16 @@ import { Modal, ModalProps } from "@mantine/core";
 import { Optional } from "../../util";
 
 /**
+ * Options for the {@link TOSModal} component.
+ */
+export type TOSModalProps = Optional<ModalProps, "onClose">;
+
+/**
  * A modal that handles open/closed state.
  *
  * Intended for use with {@link TOSPage}.
  */
-export function TOSModal({ title, onClose, ...other }: Optional<ModalProps, "onClose">) {
+export function TOSModal({ title, onClose, ...other }: TOSModalProps) {
     const [open, openHandler] = useDisclosure(true);
 
     return (